+++ views/navbar.pug
@@ -9,7 +9,7 @@ block content
             a(href="" id="menu-img"): img(src="/icons/menu.png", alt="Menu" class="icon")
             div(id="nav-search")
                 img(src="/icons/search.png" alt="Search" id="nav-search-img" class="icon")
-                input(type="search" id="nav-search-input" onsearch="searchFunction()")
+                input(type="search" id="nav-search-input")
             div(class="dropdown" img src="/icons/settings.png" alt="Settings" class="icon")
                 img(src="/icons/settings.png" alt="Settings" id="nav-settings-img" class="icon" role="button")
                 div(class="dropdown-content")
